rag-guardrails-app/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
annoy==1.17.3
anyio==4.9.0
appdirs==1.4.4
appnope==0.1.4
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
comm==0.2.2
dataclasses-json==0.6.7
datasets==3.5.1
debugpy==1.8.14
decorator==5.2.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
exceptiongroup==1.2.2
executing==2.2.0
faiss-cpu==1.11.0
fastapi==0.115.12
fastembed==0.4.0
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.6.0
fsspec==2025.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.24
langchain-community==0.3.23
langchain-core==0.3.56
langchain-openai==0.3.7
langchain-text-splitters==0.3.8
langfuse==2.60.3
langsmith==0.3.38
lark==1.2.2
litellm==1.67.4.post1
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==4.1.0
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
mypy_extensions==1.1.0
nemoguardrails==0.13.0
nest_asyncio==1.6.0
numpy==1.26.4
onnx==1.17.0
onnxruntime==1.19.2
openai==1.76.0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip==25.1
platformdirs==4.3.7
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pydantic==2.11.3
pydantic_core==2.33.1
pydantic-settings==2.9.1
Pygments==2.19.1
PyStemmer==2.2.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
ragas==0.2.15
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
safetensors==0.5.3
scikit-learn==1.6.1
sentence-transformers==4.1.0
setuptools==79.0.1
shellingham==1.5.4
simpleeval==1.0.3
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
SQLAlchemy==2.0.40
stack_data==0.6.3
starlette==0.46.2
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
typer==0.15.3
typing_extensions==4.13.2
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.2
xxhash==3.5.0
yamllint==1.37.0
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
